+++ src/Interfaces/DeliveryChargeInterface.php
@@ -3,5 +3,9 @@
 	namespace App\Interfaces;
 
 	interface DeliveryChargeInterface{
-		public function calculate(float $total_before_delivery_charges):float;
+		/**
+		* @param int $total_before_delivery_charges Basket total in cents.
+		* @return int Delivery charge in cents.
+		*/
+		public function calculate(int $total_before_delivery_charges):int;
 	}
